+++ src/shared/shared.css
@@ -927,50 +927,6 @@ span {
     pointer-events: none;
 }
 
-
-.autocomplete {
-    cursor: pointer;
-    display: inline-block;
-    position: relative;
-    height: 20px;
-    /*min-width: 100px;*/
-    margin: auto 0;
-}
-
-.autocomplete-items {
-    position: absolute;
-    max-height: 384px;
-    overflow-y: auto;
-    overflow-x: hidden;
-    top: 100%;
-    right: 0;
-    left: 12px;
-    z-index: 999;
-    margin: 0;
-    padding: 0;
-    list-style: none;
-    background-color: hsla(29, 22%, 30%, 1);
-    border: 1px solid var(--color-light-50);
-    border-top: 0px;
-}
-
-.autocomplete-items div {
-    white-space: nowrap;
-    height: 20px;
-    cursor: pointer;
-    background-color: #fff; 
-    border-bottom: 1px solid #d4d4d4; 
-}
-.autocomplete-items div:hover {
-    /*when hovering an item:*/
-    background-color: #e9e9e9; 
-}
-.autocomplete-active {
-    /*when navigating through the items using the arrow keys:*/
-    background-color: DodgerBlue !important; 
-    color: #ffffff; 
-}
-
 .actionlog {
     text-shadow: 2px 2px 2px rgb(0, 0, 0);
     white-space: nowrap;
